% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCodelistAttribute.R
\name{addCodelistAttribute}
\alias{addCodelistAttribute}
\title{Adds the cohort_codelist attribute to a cohort}
\usage{
addCodelistAttribute(cohort, codelist, cohortName = names(codelist))
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{codelist}{Named list of concepts}

\item{cohortName}{For each element of the codelist, the name of the cohort in
`cohort` to which the codelist refers}
}
\description{
Adds the cohort_codelist attribute to a cohort
}
