% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalDescriptionSpecification.R
\name{clinicalDescriptionSpecification}
\alias{clinicalDescriptionSpecification}
\title{Clinical description specification}
\usage{
clinicalDescriptionSpecification(path = NULL)
}
\arguments{
\item{path}{If NULL, specification will be returned as an R object. If a
path to a directory is provided the specification will be exported.}
}
\value{
JSON specification for clinical descriptions
}
\description{
Clinical description specification
}
