% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{clinicalTableSample}
\alias{clinicalTableSample}
\title{Helper for consistent documentation of `clinicalTableSample`.}
\arguments{
\item{clinicalTableSample}{The number of records to take a random sample for
database diagnostics. If `clinicalTableSample = NULL`, no sampling will be
performed. If `clinicalTableSample = 0` clinical table summaries will not be
run.}
}
\description{
Helper for consistent documentation of `clinicalTableSample`.
}
\keyword{internal}
