% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{cohortDoc}
\alias{cohortDoc}
\title{Helper for consistent documentation of `cohort`.}
\arguments{
\item{cohort}{Cohort table in a cdm reference}
}
\description{
Helper for consistent documentation of `cohort`.
}
\keyword{internal}
