% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortToPopulationDiagnostics.R
\name{cohortToPopulationDiagnostics}
\alias{cohortToPopulationDiagnostics}
\title{Compare characteristics of cohort matched to database population}
\usage{
cohortToPopulationDiagnostics(cohort, nSample = 1000)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{nSample}{The number of people to take a random sample for matching. If
NULL, no sampling will be performed.}
}
\value{
A summarised result
}
\description{
Compare characteristics of cohort matched to database population
}
