% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseDiagnostics.R
\name{databaseDiagnostics}
\alias{databaseDiagnostics}
\title{Database diagnostics}
\usage{
databaseDiagnostics(cohort, clinicalTableSample = NULL)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{clinicalTableSample}{The number of records to take a random sample for
database diagnostics. If `clinicalTableSample = NULL`, no sampling will be
performed. If `clinicalTableSample = 0` clinical table summaries will not be
run.}
}
\value{
A summarised result
}
\description{
phenotypeR diagnostics on the cdm object.

Diagnostics include:
* Summarise a cdm_reference object, creating a snapshot with the metadata of the cdm_reference object.
* Summarise the observation period table getting some overall statistics in a summarised_result object.
* Summarise the person table including demographics (sex, race, ethnicity, year of birth) and related statistics.
}
\examples{
\donttest{
library(omock)
library(PhenotypeR)
library(CohortConstructor)

cdm <- mockCdmFromDataset(source = "duckdb")

cdm$new_cohort <- conceptCohort(cdm,
                                conceptSet = list("codes" = c(40213201L, 4336464L)),
                                name = "new_cohort")

}
}
