% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSourceDescriptionSpecification.R
\name{dataSourceDescriptionSpecification}
\alias{dataSourceDescriptionSpecification}
\title{Data source description specification}
\usage{
dataSourceDescriptionSpecification(path = NULL)
}
\arguments{
\item{path}{If NULL, specification will be returned as an R object. If a
path to a directory is provided the specification will be exported.}
}
\value{
JSON specification for data source descriptions
}
\description{
Data source description specification
}
