% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{diagnosticsDoc}
\alias{diagnosticsDoc}
\title{Helper for consistent documentation of `diagnostics`}
\arguments{
\item{diagnostics}{A vector indicating which diagnostic analyses to perform. Options include:
`databaseDiagnostics`, `codelistDiagnostics`, `cohortDiagnostics`, and `populationDiagnostics`.}
}
\description{
Helper for consistent documentation of `diagnostics`
}
\keyword{internal}
