% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{directoryDoc}
\alias{directoryDoc}
\title{Helper for consistent documentation of `directory`.}
\arguments{
\item{directory}{Directory where to save report}
}
\description{
Helper for consistent documentation of `directory`.
}
\keyword{internal}
