% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadClinicalDescriptionTemplate.R
\name{downloadClinicalDescriptionTemplate}
\alias{downloadClinicalDescriptionTemplate}
\title{Download a Clinical Description Template}
\usage{
downloadClinicalDescriptionTemplate(
  directory,
  name = "clinical_description_template"
)
}
\arguments{
\item{directory}{Directory where to download the clinical description.}

\item{name}{Name of the Word file.Note that the file must match the cohort names
used in PhenotypeR Diagnostics if you want to integrate the clinical description
into the PhenotypeR Shiny app.}
}
\value{
A Word document with the template of the clinical description.
}
\description{
Download a Clinical Description Template
}
\examples{
\donttest{
library(PhenotypeR)
library(here)

downloadClinicalDescriptionTemplate(directory = here(),
                                    name = "metformin")


}
}
