% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabaseDescriptionTemplate.R
\name{downloadDatabaseDescriptionTemplate}
\alias{downloadDatabaseDescriptionTemplate}
\title{Download a Clinical Description Template}
\usage{
downloadDatabaseDescriptionTemplate(
  directory,
  name = "database_description_template"
)
}
\arguments{
\item{directory}{Directory where to download the database description template.}

\item{name}{Name of the Word file.Note that the file must match the database names
used in PhenotypeR Diagnostics if you want to integrate the database description
into the PhenotypeR Shiny app.}
}
\value{
A Word document with the template of the clinical description.
}
\description{
Download a Clinical Description Template
}
\examples{
\donttest{
library(PhenotypeR)
library(here)

downloadDatabaseDescriptionTemplate(directory = here(),
                                    name = "GiBleed")


}
}
