% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftClinicalDescription.R
\name{draftClinicalDescription}
\alias{draftClinicalDescription}
\title{Draft clinical descriptions using an LLM}
\usage{
draftClinicalDescription(chat, name, outputDir)
}
\arguments{
\item{chat}{An ellmer chat}

\item{name}{Clinical event of interest}

\item{outputDir}{Folder to save clinical descriptions.}
}
\value{
Creates a draft clinical description for each event of interest.
}
\description{
Draft clinical descriptions using an LLM
}
