% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{drugDiagnosticsSampleDoc}
\alias{drugDiagnosticsSampleDoc}
\title{Helper for consistent documentation of `drugDiagnosticsSample`.}
\arguments{
\item{drugDiagnosticsSample}{The number of people to take a random sample for
drug diagnostics. If `drugDiagnosticsSample = NULL`, no sampling will be
performed. If `drugDiagnosticsSample = 0` drug diagnostics will not be
run.}
}
\description{
Helper for consistent documentation of `drugDiagnosticsSample`.
}
\keyword{internal}
