% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{drugExposureSampleDoc}
\alias{drugExposureSampleDoc}
\title{Helper for consistent documentation of `drugExposureSample`.}
\arguments{
\item{drugExposureSample}{The number of people to take a random sample for
drug diagnostics. If `drugExposureSample = NULL`, no sampling will be
performed. If `drugExposureSample = 0` drug diagnostics will not be
run.}
}
\description{
Helper for consistent documentation of `drugExposureSample`.
}
\keyword{internal}
