% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{expectationsDoc}
\alias{expectationsDoc}
\title{Helper for consistent documentation of `expectations`.}
\arguments{
\item{expectations}{Data frame or tibble with cohort expectations. It must contain the following columns: cohort_name, estimate, value, and source.}
}
\description{
Helper for consistent documentation of `expectations`.
}
\keyword{internal}
