% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalDescription.R
\name{getClinicalDescription}
\alias{getClinicalDescription}
\title{Get clinical descriptions using an LLM}
\usage{
getClinicalDescription(chat, name, outputDir)
}
\arguments{
\item{chat}{An ellmer chat}

\item{name}{Clinical event of interest}

\item{outputDir}{Folder to save clinical descriptions.}
}
\value{
A list with each item containing the clinical description.
}
\description{
Get clinical descriptions using an LLM
}
