% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortExpectations.R
\name{getCohortExpectations}
\alias{getCohortExpectations}
\title{Get cohort expectations using an LLM}
\usage{
getCohortExpectations(chat, phenotypes)
}
\arguments{
\item{chat}{An ellmer chat}

\item{phenotypes}{Either a vector of phenotype names or results from
PhenotypeR.}
}
\value{
A tibble with expectations about the cohort.
}
\description{
Get cohort expectations using an LLM
}
