% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftClinicalDescription.R
\name{importClinicalDescription}
\alias{importClinicalDescription}
\title{Import clinical descriptions}
\usage{
importClinicalDescription(path)
}
\arguments{
\item{path}{Either a directory containing clinical descriptions or a path to
a specific clinical description}
}
\value{
A list of clinical descriptions
}
\description{
Import clinical descriptions
}
