% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{matchedSampleDoc}
\alias{matchedSampleDoc}
\title{Helper for consistent documentation of `matchedSample`.}
\arguments{
\item{matchedSample}{The number of people to take a random sample for
matching. If NULL, no sampling will be performed.}
}
\description{
Helper for consistent documentation of `matchedSample`.
}
\keyword{internal}
