% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementDiagnostics.R
\name{measurementDiagnostics}
\alias{measurementDiagnostics}
\title{Diagnostics of a conceptSet of measurement codes}
\usage{
measurementDiagnostics(cdm, codes, cohort = NULL, timing = "during")
}
\arguments{
\item{cdm}{A reference to the cdm object.}

\item{codes}{A codelist of measurement codes for which to perform diagnostics.}

\item{cohort}{A cohort in which to perfom the diagnostics of the measurement
codes provided. If NULL, the cohort will be set to everyone in observation in
the database.}

\item{timing}{Three options: 1) "any" if the interest is on measurement
recorded any time, 2) "during", if interested in measurements while the
subject is in the cohort (or in observation if cohort = NULL), and 3)
"cohort_start_date" for measurements ocurring at cohort start date (or at
"observation_period_start_date if cohort = NULL).}
}
\value{
A summarised result
}
\description{
Diagnostics of a conceptSet of measurement codes
}
\examples{
\donttest{
library(PhenotypeR)

cdm <- mockPhenotypeR()

# diagnostics in the database for measurements occurring while patients are
# in observation
result <- measurementDiagnostics(
  cdm = cdm, codes = list("test_codelist" = c(3001467L, 45875977L)),
  timing = "during"
)

# diagnostics subsetted to "my_cohort" for measurements occurring at cohort
# start date
result_subset <- measurementDiagnostics(
  cdm = cdm, codes = list("test_codelist" = c(3001467L, 45875977L)),
  cohort = cdm$my_cohort, timing = "cohort_start_date"
)

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
