% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{measurementDiagnosticsSampleDoc}
\alias{measurementDiagnosticsSampleDoc}
\title{Helper for consistent documentation of \code{measurementDiagnosticsSample}.}
\arguments{
\item{measurementDiagnosticsSample}{The number of people to take a random sample for
measurement diagnostics. If \code{measurementDiagnosticsSample = NULL}, no sampling will be
performed. If \code{measurementDiagnosticsSample = 0} measurement diagnostics will not be
run.}
}
\description{
Helper for consistent documentation of \code{measurementDiagnosticsSample}.
}
\keyword{internal}
