% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{measurementSampleDoc}
\alias{measurementSampleDoc}
\title{Helper for consistent documentation of `measurementSample`.}
\arguments{
\item{measurementSample}{The number of people to take a random sample for
measurement diagnostics. If `measurementSample = NULL`, no sampling will be
performed. If `measurementSample = 0` measurement diagnostics will not be
run.}
}
\description{
Helper for consistent documentation of `measurementSample`.
}
\keyword{internal}
