% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeCohort.R
\name{phenotypeCohort}
\alias{phenotypeCohort}
\title{Phenotype a set of cohorts}
\usage{
phenotypeCohort(cohort)
}
\arguments{
\item{cohort}{Cohort table}
}
\value{
A summarised result
}
\description{
Phenotype a set of cohorts
}
