% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{populationSampleDoc}
\alias{populationSampleDoc}
\title{Helper for consistent documentation of `populationSample`.}
\arguments{
\item{populationSample}{Number of people from the cdm to sample. If NULL no
sampling will be performed}
}
\description{
Helper for consistent documentation of `populationSample`.
}
\keyword{internal}
