% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{suppress}
\alias{settings}
\alias{bind}
\alias{importSummarisedResult}
\alias{exportSummarisedResult}
\alias{summariseAchillesCodeUse}
\alias{summariseCohortCodeUse}
\alias{summariseCodeUse}
\alias{summariseOrphanCodes}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CodelistGenerator}{\code{\link[CodelistGenerator]{summariseAchillesCodeUse}}, \code{\link[CodelistGenerator]{summariseCodeUse}}, \code{\link[CodelistGenerator]{summariseCohortCodeUse}}, \code{\link[CodelistGenerator]{summariseOrphanCodes}}}

  \item{omopgenerics}{\code{\link[omopgenerics]{bind}}, \code{\link[omopgenerics]{exportSummarisedResult}}, \code{\link[omopgenerics]{importSummarisedResult}}, \code{\link[omopgenerics]{settings}}, \code{\link[omopgenerics]{suppress}}}
}}

