% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPhenotype.R
\name{reportPhenotype}
\alias{reportPhenotype}
\title{Create a report summarising your phenotyping results}
\usage{
reportPhenotype(result, directory = here::here())
}
\arguments{
\item{result}{A summarised result}

\item{directory}{Directory where to save report}
}
\value{
A report
}
\description{
Create a report summarising your phenotyping results
}
