% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{resultDoc}
\alias{resultDoc}
\title{Helper for consistent documentation of `result`.}
\arguments{
\item{result}{A summarised result}
}
\description{
Helper for consistent documentation of `result`.
}
\keyword{internal}
