% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDescriptions.R
\name{shinyClinicalDescriptions}
\alias{shinyClinicalDescriptions}
\title{Shiny app to create clinical descriptions for contextualising diagnostic results}
\usage{
shinyClinicalDescriptions(directory, open = rlang::is_interactive())
}
\arguments{
\item{directory}{Directory where to save shiny app}

\item{open}{If TRUE, the shiny app will be launched in a new session. If
FALSE, the shiny app will be created but not launched.}
}
\value{
Shiny app
}
\description{
Shiny app to create clinical descriptions for contextualising diagnostic results
}
\examples{
\donttest{
shinyClinicalDescriptions(tempdir())
}
}
