% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPhenotype.R
\name{shinyPhenotype}
\alias{shinyPhenotype}
\title{Create a shiny app summarising your phenotyping results}
\usage{
shinyPhenotype(result)
}
\arguments{
\item{result}{A summarised result}
}
\value{
A shiny app
}
\description{
Create a shiny app summarising your phenotyping results
}
