% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugDiagnostics.R
\name{summariseDrugExposureDiagnostics}
\alias{summariseDrugExposureDiagnostics}
\title{Summarise diagnostics of the drug exposure table for a given concept set
and/or ingredient}
\usage{
summariseDrugExposureDiagnostics(
  cdm,
  ingredient = NULL,
  conceptSet = NULL,
  checks = c("missing", "exposureDuration", "quantity"),
  sample = 10000,
  dateRange = c("2010-01-01", NA),
  byConcept = FALSE
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{ingredient}{An ingredient concept ID.}

\item{conceptSet}{A concept set.}

\item{checks}{the checks to be executed, by default the missing values, the
exposure duration and the quantity. Possible options are "missing",
"exposureDuration", "type", "route", "sourceConcept", "daysSupply",
"verbatimEndDate", "dose", "sig", "quantity", "daysBetween" and
"diagnosticsSummary". Note "standardConcept" check is always performed.}

\item{sample}{Number of records to use for the checks. Note 'overall' and
'sourceConcept' checks will be performed with the}

\item{dateRange}{Range to use the records}

\item{byConcept}{Whether to stratify results by concept id.}
}
\value{
A summarised_result object with the checks.
}
\description{
Summarise diagnostics of the drug exposure table for a given concept set
and/or ingredient
}
\examples{
\donttest{
library(PhenotypeR)
library(omock)

cdm <- mockCdmFromDataset()

summariseDrugExposureDiagnostics(cdm = cdm, ingredient = 1125315L)
}

}
