% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugDiagnostics.R
\name{tableDrugExposureDiagnostics}
\alias{tableDrugExposureDiagnostics}
\title{Visusalise the results in a table object}
\usage{
tableDrugExposureDiagnostics(
  result,
  check = "standardConcept",
  header = NULL,
  groupColumn = NULL,
  hide = NULL,
  type = NULL,
  style = NULL
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{check}{The check to visualise in the table.}

\item{header}{A vector specifying the elements to include in the header. The
order of elements matters, with the first being the topmost header. Elements
in header can be:
- Any of the columns returned by tableColumns(result) to create a header for
these columns.
- Any other input to create an overall header.}

\item{groupColumn}{Columns to use as group labels, to see options use
tableColumns(result). By default, the name of the new group will be the tidy*
column names separated by ";". To specify a custom group name, use a named
list such as: list("newGroupName" = c("variable_name", "variable_level")).
*tidy: The tidy format applied to column names replaces "_" with a space an
converts to sentence case. Use rename to customise specific column names.}

\item{hide}{Columns to drop from the output table.}

\item{type}{Character string specifying the desired output table format. See
`visOmopResults::tableType()` for supported table types. If `type = NULL`,
global options (set via `visOmopResults::setGlobalTableOptions()`) will be
used if available; otherwise, a default 'gt' table is created.}

\item{style}{Defines the visual formatting of the table. This argument can be
provided in one of the following ways:
1. **Pre-defined style**: Use the name of a built-in style (e.g., "darwin").
See tableStyle() for available options.
2. **YAML file path**: Provide the path to an existing .yml file defining a
new style.
3. **List of custome R code**: Supply a block of custom R code or a named
list describing styles for each table section. This code must be specific to
the selected table type. If `style = NULL`, the function will use global
options (see `visOmopResults::setGlobalTableOptions()`) or an existing
`_brand.yml` file (if found); otherwise, the default style is applied. For
more details, see the Styles vignette on the package website.}
}
\value{
A table visualisation.
}
\description{
Visusalise the results in a table object
}
\examples{
\donttest{
library(PhenotypeR)
library(omock)

cdm <- mockCdmFromDataset()

result <- summariseDrugExposureDiagnostics(cdm = cdm, ingredient = 1125315L)

tableDrugExposureDiagnostics(result = result)
}

}
