% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortExpectations.R
\name{viewCohortExpectations}
\alias{viewCohortExpectations}
\title{View cohort expectations}
\usage{
viewCohortExpectations(expectations, type = "gt")
}
\arguments{
\item{expectations}{Data frame or tibble with cohort expectations}

\item{type}{Table type to view results. See visOmopResults::tableType()
for supported tables.}
}
\value{
Summary of cohort expectations
}
\description{
View cohort expectations
}
