% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelistDiagnostics.R
\name{codelistDiagnostics}
\alias{codelistDiagnostics}
\title{Run codelist-level diagnostics}
\usage{
codelistDiagnostics(cohort)
}
\arguments{
\item{cohort}{Cohort table}
}
\value{
A summarised result
}
\description{
Run codelist-level diagnostics
}
