% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelistDiagnostics.R
\name{codelistDiagnostics}
\alias{codelistDiagnostics}
\title{Run codelist-level diagnostics}
\usage{
codelistDiagnostics(cohort)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference. The cohort_codelist
attribute must be populated. The cdm reference must contain achilles
tables as these will be used for deriving concept counts.}
}
\value{
A summarised result
}
\description{
Run codelist-level diagnostics
}
