% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnostics.R
\name{cohortDiagnostics}
\alias{cohortDiagnostics}
\title{Run cohort-level diagnostics}
\usage{
cohortDiagnostics(cohort)
}
\arguments{
\item{cohort}{Cohort table}
}
\value{
A summarised result
}
\description{
Run cohort-level diagnostics
}
