% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnostics.R
\name{cohortDiagnostics}
\alias{cohortDiagnostics}
\title{Run cohort-level diagnostics}
\usage{
cohortDiagnostics(cohort, strata = list())
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{strata}{A list of variables to stratify results. These variables must
have been added as additional columns in the cohort table.}
}
\value{
A summarised result
}
\description{
Run cohort-level diagnostics
}
