% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnostics.R
\name{cohortDiagnostics}
\alias{cohortDiagnostics}
\title{Run cohort-level diagnostics}
\usage{
cohortDiagnostics(cohort, strata = list(), matchCohort = TRUE, nSample = NULL)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{strata}{A list of variables to stratify results. These variables must
have been added as additional columns in the cohort table.}

\item{matchCohort}{TRUE or FALSE. If TRUE age and sex matched cohorts will
be added for each cohort in the cohort table}

\item{nSample}{The number of people to take a random sample for running
large scale characteristics}
}
\value{
A summarised result
}
\description{
Run cohort-level diagnostics
}
