% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnostics.R
\name{cohortDiagnostics}
\alias{cohortDiagnostics}
\title{Run cohort-level diagnostics}
\usage{
cohortDiagnostics(cohort, strata = list())
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{strata}{A list of variables to stratify results. These variables must
have been added as additional columns in the cohort table.}
}
\value{
A summarised result
}
\description{
Run cohort-level diagnostics
}
\examples{
\donttest{
  cdm_local <- omock::mockCdmReference() |>
   omock::mockPerson(nPerson = 100) |>
   omock::mockObservationPeriod() |>
   omock::mockConditionOccurrence() |>
   omock::mockDrugExposure() |>
   omock::mockObservation() |>
   omock::mockMeasurement() |>
   omock::mockCohort(name = "my_cohort")
   cdm_local$visit_occurrence <- dplyr::tibble(
    person_id = 1L,
    visit_occurrence_id = 1L,
    visit_concept_id = 1L,
    visit_start_date = as.Date("2000-01-01"),
    visit_end_date = as.Date("2000-01-01"),
    visit_type_concept_id = 1L
    )
    cdm_local$procedure_occurrence <- dplyr::tibble(
      person_id = 1L,
      procedure_occurrence_id = 1L,
      procedure_concept_id = 1L,
      procedure_date = as.Date("2000-01-01"),
      procedure_type_concept_id = 1L
      )
 db <- DBI::dbConnect(duckdb::duckdb())
 cdm <- CDMConnector::copyCdmTo(con = db,
                                cdm = cdm_local,
                                schema ="main",
                                overwrite = TRUE)

 cdm$my_cohort |> cohortDiagnostics()
 CDMConnector::cdmDisconnect(cdm = cdm)
}
}
