% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortDiagnostics.R
\name{cohortDiagnostics}
\alias{cohortDiagnostics}
\title{Run cohort-level diagnostics}
\usage{
cohortDiagnostics(cohort)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}
}
\value{
A summarised result
}
\description{
Runs phenotypeR diagnostics on the cohort.
The diganotics include:
* Age groups and sex summarised.
* A summary of visits of everyone in the cohort using visit_occurrence table.
* A summary of age and sex density of the cohort.
* Attritions of the cohorts.
* Overlap between cohorts (if more than one cohort is being used).
}
\examples{
\donttest{
  cdm_local <- omock::mockCdmReference() |>
   omock::mockPerson(nPerson = 100) |>
   omock::mockObservationPeriod() |>
   omock::mockConditionOccurrence() |>
   omock::mockDrugExposure() |>
   omock::mockObservation() |>
   omock::mockMeasurement() |>
   omock::mockCohort(name = "my_cohort")
   cdm_local$visit_occurrence <- dplyr::tibble(
    person_id = 1L,
    visit_occurrence_id = 1L,
    visit_concept_id = 1L,
    visit_start_date = as.Date("2000-01-01"),
    visit_end_date = as.Date("2000-01-01"),
    visit_type_concept_id = 1L
    )
    cdm_local$procedure_occurrence <- dplyr::tibble(
      person_id = 1L,
      procedure_occurrence_id = 1L,
      procedure_concept_id = 1L,
      procedure_date = as.Date("2000-01-01"),
      procedure_type_concept_id = 1L
      )
 db <- DBI::dbConnect(duckdb::duckdb())
 cdm <- CDMConnector::copyCdmTo(con = db,
                                cdm = cdm_local,
                                schema ="main",
                                overwrite = TRUE)

 cdm$my_cohort |> cohortDiagnostics()
 CDMConnector::cdmDisconnect(cdm = cdm)
}
}
