% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{cohortSampleDoc}
\alias{cohortSampleDoc}
\title{Helper for consistent documentation of `cohortSample`.}
\arguments{
\item{cohortSample}{The number of people to take a random sample for cohortDiagnostics. If `cohortSample = NULL`, no sampling will be performed.}
}
\description{
Helper for consistent documentation of `cohortSample`.
}
\keyword{internal}
