% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalDescription.R
\name{importDatabaseDescription}
\alias{importDatabaseDescription}
\title{Import database descriptions}
\usage{
importDatabaseDescription(path)
}
\arguments{
\item{path}{Either a directory containing database descriptions or a path to
a specific database description}
}
\value{
A list of database descriptions
}
\description{
Import database descriptions
}
