% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchedDiagnostics.R
\name{matchedDiagnostics}
\alias{matchedDiagnostics}
\title{Compare characteristics of cohort matched to database population}
\usage{
matchedDiagnostics(cohort, matchedSample = 1000)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{matchedSample}{The number of people to take a random sample for
matching. If NULL, no sampling will be performed.}
}
\value{
A summarised result
}
\description{
Compare characteristics of cohort matched to database population
}
\examples{
\donttest{
  cdm_local <- omock::mockCdmReference() |>
    omock::mockPerson(nPerson = 100) |>
    omock::mockObservationPeriod() |>
    omock::mockConditionOccurrence() |>
    omock::mockDrugExposure() |>
    omock::mockObservation() |>
    omock::mockMeasurement() |>
    omock::mockCohort(name = "my_cohort")
 cdm_local$visit_occurrence <- dplyr::tibble(
   person_id = 1L,
   visit_occurrence_id = 1L,
   visit_concept_id = 1L,
   visit_start_date = as.Date("2000-01-01"),
   visit_end_date = as.Date("2000-01-01"),
   visit_type_concept_id = 1L)
 cdm_local$procedure_occurrence <- dplyr::tibble(
   person_id = 1L,
   procedure_occurrence_id = 1L,
   procedure_concept_id = 1L,
   procedure_date = as.Date("2000-01-01"),
   procedure_type_concept_id = 1L)

 db <- DBI::dbConnect(duckdb::duckdb())
 cdm <- CDMConnector::copyCdmTo(con = db,
                                cdm = cdm_local,
                                schema ="main",
                                overwrite = TRUE)
 result <- cdm$my_cohort |> matchedDiagnostics()
 CDMConnector::cdm_disconnect(cdm)
}
}
