% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{matchedDoc}
\alias{matchedDoc}
\title{Helper for consistent documentation of `matched` and `matchedSample`.}
\arguments{
\item{matchedAnalysis}{Boolean variable. Whether to conduct the analysis for the matched cohorts (TRUE) or not (FALSE).}

\item{matchedSample}{Only if matchedAnalysis = TRUE. The number of people to take a random sample for
matching. If NULL, no sampling will be performed.}
}
\description{
Helper for consistent documentation of `matched` and `matchedSample`.
}
\keyword{internal}
