% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{measurementDiagnosticsSampleDoc}
\alias{measurementDiagnosticsSampleDoc}
\title{Helper for consistent documentation of `measurementDiagnosticsSample`.}
\arguments{
\item{measurementDiagnosticsSample}{The number of people to take a random sample for
measurement diagnostics. If `measurementDiagnosticsSample = NULL`, no sampling will be
performed. If `measurementDiagnosticsSample = 0` measurement diagnostics will not be
run.}
}
\description{
Helper for consistent documentation of `measurementDiagnosticsSample`.
}
\keyword{internal}
