% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Phenotype a cohort}
\usage{
phenotype(cohort, nSample = 1000)
}
\arguments{
\item{cohort}{Cohort}

\item{nSample}{The number of people to take a random sample for matching to
the database population. If NULL, no sampling will be performed and the
entire cohorts will be used.}
}
\value{
A summarised result
}
\description{
Phenotype a cohort
}
