% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeDiagnostics.R
\name{phenotypeDiagnostics}
\alias{phenotypeDiagnostics}
\title{Phenotype a cohort}
\usage{
phenotypeDiagnostics(
  cohort,
  databaseDiagnostics = list(),
  codelistDiagnostics = list(),
  cohortDiagnostics = list(),
  populationDiagnostics = list()
)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{databaseDiagnostics}{A list of arguments that uses `databaseDiagnostics`.
If the list is empty, the default values will be used.
Example:
*databaseDiagnostics = list(
"diagnostics" = c("snapshot", "person", "observationPeriods", "clinicalRecords")
 )*}

\item{codelistDiagnostics}{A list of arguments that uses `codelistDiagnostics`.
If the list is empty, the default values will be used.
Example:
*codelistDiagnostics = list(
 "diagnostics" = c("achillesCodeUse", "orphanCodeUse", "cohortCodeUse",
 "drugDiagnostics", "measurementDiagnostics"),
 "measurementDiagnosticsSample" = 20000,
 "drugDiagnosticsSample" = 20000
  )*}

\item{cohortDiagnostics}{A list of arguments that uses `cohortDiagnostics`.
If the list is empty,
the default values will be used.
Example:
*cohortDiagnostics = list(
 "diagnostics" = c("cohortCount", "cohortCharacteristics", "largeScaleCharacteristics",
                   "compareCohorts", "cohortSurvival),
 "cohortSample" = 20000,
 "matchedSample" = 1000
 )*}

\item{populationDiagnostics}{A list of arguments that uses `populationDiagnostics`.
If the list is empty, the default values will be used.
Example:
*populationDiagnostics = list(
 "diagnostics" = c("incidence", "periodPrevalence"),
 "populationSample" = 1000000,
 "populationDateRange" = as.Date(c(NA,NA))
 )*}
}
\value{
A summarised result
}
\description{
This comprises all the diagnostics that are being offered in this package,
this includes:

* A diagnostics on the database via `databaseDiagnostics`.
* A diagnostics on the cohort_codelist attribute of the cohort via `codelistDiagnostics`.
* A diagnostics on the cohort via `cohortDiagnostics`.
* A diagnostics on the population via `populationDiagnostics`.
}
\examples{
\donttest{
library(omock)
library(CohortConstructor)
library(PhenotypeR)

cdm <- mockCdmFromDataset(source = "duckdb")
cdm$warfarin <- conceptCohort(cdm,
                              conceptSet =  list(warfarin = c(1310149L,
                                                              40163554L)),
                              name = "warfarin")

# Run PhenotypeR with the default values. If you want to check which are the
# default values, use:
# `formals(populationDiagnostics)`
result <- phenotypeDiagnostics(cdm$warfarin)

# Notice that the previous line of code will give the same results as typing manually
# all the default values:
result <- phenotypeDiagnostics(cdm$warfarin,
                               databaseDiagnostics = list(
                                 "diagnostics" = c("snapshot", "person",
                                 "observationPeriods", "clinicalRecords"),
                               ),
                               codelistDiagnostics = list(
                                 "diagnostics" = c("achillesCodeUse", "orphanCodeUse",
                                                   "cohortCodeUse", "drugDiagnostics",
                                                   "measurementDiagnostics"),
                                 "measurementDiagnosticsSample" = 20000,
                                 "drugDiagnosticsSample" = 20000
                               ),
                               cohortDiagnostics = list(
                                 "diagnostics" = c("cohortCount", "cohortCharacteristics",
                                                   "largeScaleCharacteristics",
                                                   "compareCohorts"),
                                 "cohortSample" = 20000,
                                 "matchedSample" = 1000
                               ),
                               populationDiagnostics = list(
                                 "diagnostics" = c("incidence", "periodPrevalence"),
                                 "populationSample" = 1000000,
                                 "populationDateRange" = as.Date(c(NA,NA))
                               ))

# By default, cohortSurvival analysis will not be run. If you want to run it, please use:
# result <- phenotypeDiagnostics(cdm$warfarin,
                               cohortDiagnostics = list(
                                 "diagnostics" = c("cohortCount", "cohortCharacteristics",
                                                   "largeScaleCharacteristics",
                                                   "compareCohorts", "cohortSurvival")))


# Run PhenotypeR with the default values, except for populationSample:
result <- phenotypeDiagnostics(cdm$warfarin,
                               populationDiagnostics = list("populationSample" = 1000))
}
}
