% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeDiagnostics.R
\name{phenotypeDiagnostics}
\alias{phenotypeDiagnostics}
\title{Phenotype a cohort}
\usage{
phenotypeDiagnostics(
  cohort,
  databaseDiagnostics = TRUE,
  codelistDiagnostics = TRUE,
  cohortDiagnostics = TRUE,
  matchedDiagnostics = TRUE,
  nSample = 1000
)
}
\arguments{
\item{cohort}{Cohort}

\item{databaseDiagnostics}{If TRUE, database diagnostics will be run.}

\item{codelistDiagnostics}{If TRUE, codelist diagnostics will be run.}

\item{cohortDiagnostics}{If TRUE, cohort diagnostics will be run.}

\item{matchedDiagnostics}{If TRUE, cohort to population
diagnostics will be run.}

\item{nSample}{The number of people to take a random sample for matching to
the database population. If NULL, no sampling will be performed and the
entire cohorts will be used.}
}
\value{
A summarised result
}
\description{
Phenotype a cohort
}
