% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeR-package.R
\docType{package}
\name{phenotypeR-package}
\alias{phenotypeR}
\alias{phenotypeR-package}
\title{phenotypeR: Phenotype a set of cohorts in data mapped to a the OMOP common data model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
