% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationDiagnostics.R
\name{populationDiagnostics}
\alias{populationDiagnostics}
\title{Population-level diagnostics}
\usage{
populationDiagnostics(cohort, populationSample = 1e+06)
}
\arguments{
\item{cohort}{Cohort table in a cdm reference}

\item{populationSample}{Number of people from the cdm to sample. If NULL no
sampling will be performed}
}
\value{
A summarised result
}
\description{
Population-level diagnostics
}
\examples{
\donttest{
library(IncidencePrevalence)
cdm <- mockIncidencePrevalenceRef(sampleSize = 1000)
cdm <- generateDenominatorCohortSet(cdm, name = "denom")
pop_diag <- populationDiagnostics(cohort = cdm$outcome,
                                  populationSample = 250)
CDMConnector::cdm_disconnect(cdm)
}
}
