% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationDiagnostics.R
\name{populationDiagnostics}
\alias{populationDiagnostics}
\title{Population-level diagnostics}
\usage{
populationDiagnostics(cohort, populationSample = 1e+06)
}
\arguments{
\item{cohort}{Cohort}

\item{populationSample}{N of people from the cdm to sample. If NULL no
sampling will be performed}
}
\description{
Population-level diagnostics
}
