% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDescriptions.R
\name{shinyDescriptions}
\alias{shinyDescriptions}
\title{Create a shiny app creating clinical and database descriptions}
\usage{
shinyDescriptions(directory, open = rlang::is_interactive())
}
\arguments{
\item{directory}{Directory where to save report}

\item{open}{If TRUE, the shiny app will be launched in a new session. If
FALSE, the shiny app will be created but not launched.}
}
\value{
A shiny app
}
\description{
A shiny app that is designed for creating descriptions for the PhenotrpeR
dianostics shiny app.
}
\examples{
\donttest{
library(PhenotypeR)
shinyDescriptions(tempdir())
}
}
