% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDiagnostics.R
\name{shinyDiagnostics}
\alias{shinyDiagnostics}
\title{Create a shiny app summarising your phenotyping results}
\usage{
shinyDiagnostics(result)
}
\arguments{
\item{result}{A summarised result}
}
\value{
A shiny app
}
\description{
Create a shiny app summarising your phenotyping results
}
