% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelper.R
\name{survivalDoc}
\alias{survivalDoc}
\title{Helper for consistent documentation of `survival`.}
\arguments{
\item{survival}{Boolean variable. Whether to conduct survival analysis (TRUE) or not (FALSE).}
}
\description{
Helper for consistent documentation of `survival`.
}
\keyword{internal}
